refactor: chat 줄바꿈 기능 추가
