+++ packages/gdsc-dju-admin/src/components/molecules/ApplicantChatSection/styled.ts
@@ -1,28 +1,30 @@
 import styled, { css } from 'styled-components';
 
 export const ApplicantChatSectionWrapper = styled.div`
-  width: 400px;
-  height: 100%;
+  width: 100%;
   display: flex;
+  height: calc(100vh - 40px);
   flex-direction: column;
   position: relative;
+  border-left: 1px solid ${({ theme }) => theme.colors.grey200};
 `;
 export const ApplicantChatBottomBar = styled.div`
-  position: absolute;
   bottom: 0;
   padding: 10px;
   width: 100%;
+  height: fit-content;
   box-sizing: border-box;
   display: flex;
   flex-direction: row;
 `;
-export const ApplicantChatInput = styled.input`
+export const ApplicantChatInput = styled.textarea`
   border: 1px solid ${({ theme }) => theme.colors.grey300};
+  font-size: ${({ theme }) => theme.fontSize.body2};
   border-radius: 7px;
-  height: 36px;
+  height: fit-content;
   width: 100%;
   box-sizing: border-box;
-  padding: 0 10px;
+  padding: 6px 10px;
   margin-right: 5px;
 `;
 export const ApplicantChatSendButton = styled.button`
@@ -34,7 +36,7 @@ export const ApplicantChatSendButton = styled.button`
 `;
 export const ApplicantChatList = styled.div`
   width: 100%;
-  height: 700px;
+  height: 100%;
   overflow-y: scroll;
   position: relative;
   padding: 10px;
