+++ packages/gdsc-dju-admin/src/components/molecules/ApplicantInfoState/styled.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const ApplicantInfoStateWrapper = styled.section`
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  row-gap: 10px;
+  gap: 10px;
+`;
+export const ApplicantBadgeWrapper = styled.div`
+  display: flex;
+  height: fit-content;
+`;
