+++ packages/gdsc-dju-admin/src/components/molecules/Applicantion/styled.ts
@@ -1,0 +1,20 @@
+import { motion } from 'framer-motion';
+import styled from 'styled-components';
+
+export const ApplicationWrapper = styled.div`
+  width: 100%;
+  padding: 20px;
+  height: 100%;
+  box-sizing: border-box;
+  overflow: scroll;
+`;
+export const ApplicationHeader = styled.div`
+  font-size: ${({ theme }) => theme.fontSize.h7};
+  color: ${({ theme }) => theme.colors.grey900};
+  margin-bottom: 6px;
+`;
+export const ApplicationText = styled(motion.p)`
+  font-size: ${({ theme }) => theme.fontSize.body2};
+  color: ${({ theme }) => theme.colors.grey800};
+  margin-bottom: 20px;
+`;
