+++ packages/gdsc-dju-admin/src/components/molecules/modal/AdminEmailCheckModal/index.tsx
@@ -15,7 +15,7 @@ import {
 } from '../styled';
 
 interface Props {
-  sendEmail: (
+  emailCheckHandler: (
     template: string | null,
     applicants: IApplicantTypeWithID[],
   ) => void;
@@ -24,7 +24,7 @@ interface Props {
 }
 
 const AdminEmailCheckModal: React.FC<Props> = ({
-  sendEmail,
+  emailCheckHandler,
   template,
   applicants,
 }) => {
@@ -55,7 +55,7 @@ const AdminEmailCheckModal: React.FC<Props> = ({
                 <GDSCButton
                   text={'전송하기'}
                   color={'red900'}
-                  onClick={() => sendEmail(template, applicants)}
+                  onClick={() => emailCheckHandler(template, applicants)}
                 />
               </ButtonWrapper>
             </ModalInner>
