+++ packages/gdsc-dju-admin/src/pages/index.tsx
@@ -15,7 +15,8 @@ const Admin = () => {
       <AdminContainer>
         <Routes>
           <Route path={'/*'} element={<Home />} />
-          <Route path={'/recruit'} element={<AdminApplicants />} />
+          <Route path={'/recruit/*'} element={<AdminApplicants />} />
+          <Route path={'/recruit/:userid'} element={<AdminApplicants />} />
           <Route path={'/email'} element={<Email />} />
           <Route path={'/email-log'} element={<AdminEmailLog />} />
         </Routes>
